+++ packages/wallet-connectors/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Added exported constant `CONCORDIUM_WALLET_CONNECT_PROJECT_ID` that dApps should use when desiring to connect to a Concordium mobile wallet.
+-   Added exported constant `CONCORDIUM_WALLET_CONNECT_PROJECT_ID` that dApps may use when connecting to a Concordium mobile wallet.
 
 ### Changed