+++ packages/wallet-connectors/CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   `WalletConnection`: `signMessage` now supports signing arbitrary data.
+
+## [0.2.3] - 2023-04-03
+
+No changes. Had to bump version to fix NPM release.
+
 ## [0.2.2] - 2023-04-03
 
 ### Added
 
--   Adds support for signing arbitrary data with signMessage.
 -   Added exported constant `CONCORDIUM_WALLET_CONNECT_PROJECT_ID` that dApps may use when connecting to a Concordium mobile wallet.
 
 ### Changed